+++ packages/api_client/lib/src/models/catalog_item_dto.freezed.dart
@@ -1,0 +1,283 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// coverage:ignore-file
+// ignore_for_file: type=lint
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
+
+part of 'catalog_item_dto.dart';
+
+// **************************************************************************
+// FreezedGenerator
+// **************************************************************************
+
+// dart format off
+T _$identity<T>(T value) => value;
+
+/// @nodoc
+mixin _$CatalogItemDto {
+
+ String get id; String get name; String get description;
+/// Create a copy of CatalogItemDto
+/// with the given fields replaced by the non-null parameter values.
+@JsonKey(includeFromJson: false, includeToJson: false)
+@pragma('vm:prefer-inline')
+$CatalogItemDtoCopyWith<CatalogItemDto> get copyWith => _$CatalogItemDtoCopyWithImpl<CatalogItemDto>(this as CatalogItemDto, _$identity);
+
+  /// Serializes this CatalogItemDto to a JSON map.
+  Map<String, dynamic> toJson();
+
+
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is CatalogItemDto&&(identical(other.id, id) || other.id == id)&&(identical(other.name, name) || other.name == name)&&(identical(other.description, description) || other.description == description));
+}
+
+@JsonKey(includeFromJson: false, includeToJson: false)
+@override
+int get hashCode => Object.hash(runtimeType,id,name,description);
+
+@override
+String toString() {
+  return 'CatalogItemDto(id: $id, name: $name, description: $description)';
+}
+
+
+}
+
+/// @nodoc
+abstract mixin class $CatalogItemDtoCopyWith<$Res>  {
+  factory $CatalogItemDtoCopyWith(CatalogItemDto value, $Res Function(CatalogItemDto) _then) = _$CatalogItemDtoCopyWithImpl;
+@useResult
+$Res call({
+ String id, String name, String description
+});
+
+
+
+
+}
+/// @nodoc
+class _$CatalogItemDtoCopyWithImpl<$Res>
+    implements $CatalogItemDtoCopyWith<$Res> {
+  _$CatalogItemDtoCopyWithImpl(this._self, this._then);
+
+  final CatalogItemDto _self;
+  final $Res Function(CatalogItemDto) _then;
+
+/// Create a copy of CatalogItemDto
+/// with the given fields replaced by the non-null parameter values.
+@pragma('vm:prefer-inline') @override $Res call({Object? id = null,Object? name = null,Object? description = null,}) {
+  return _then(_self.copyWith(
+id: null == id ? _self.id : id // ignore: cast_nullable_to_non_nullable
+as String,name: null == name ? _self.name : name // ignore: cast_nullable_to_non_nullable
+as String,description: null == description ? _self.description : description // ignore: cast_nullable_to_non_nullable
+as String,
+  ));
+}
+
+}
+
+
+/// Adds pattern-matching-related methods to [CatalogItemDto].
+extension CatalogItemDtoPatterns on CatalogItemDto {
+/// A variant of `map` that fallback to returning `orElse`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
+
+@optionalTypeArgs TResult maybeMap<TResult extends Object?>(TResult Function( _CatalogItemDto value)?  $default,{required TResult orElse(),}){
+final _that = this;
+switch (_that) {
+case _CatalogItemDto() when $default != null:
+return $default(_that);case _:
+  return orElse();
+
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// Callbacks receives the raw object, upcasted.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case final Subclass2 value:
+///     return ...;
+/// }
+/// ```
+
+@optionalTypeArgs TResult map<TResult extends Object?>(TResult Function( _CatalogItemDto value)  $default,){
+final _that = this;
+switch (_that) {
+case _CatalogItemDto():
+return $default(_that);case _:
+  throw StateError('Unexpected subclass');
+
+}
+}
+/// A variant of `map` that fallback to returning `null`.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case final Subclass value:
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
+
+@optionalTypeArgs TResult? mapOrNull<TResult extends Object?>(TResult? Function( _CatalogItemDto value)?  $default,){
+final _that = this;
+switch (_that) {
+case _CatalogItemDto() when $default != null:
+return $default(_that);case _:
+  return null;
+
+}
+}
+/// A variant of `when` that fallback to an `orElse` callback.
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return orElse();
+/// }
+/// ```
+
+@optionalTypeArgs TResult maybeWhen<TResult extends Object?>(TResult Function( String id,  String name,  String description)?  $default,{required TResult orElse(),}) {final _that = this;
+switch (_that) {
+case _CatalogItemDto() when $default != null:
+return $default(_that.id,_that.name,_that.description);case _:
+  return orElse();
+
+}
+}
+/// A `switch`-like method, using callbacks.
+///
+/// As opposed to `map`, this offers destructuring.
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case Subclass2(:final field2):
+///     return ...;
+/// }
+/// ```
+
+@optionalTypeArgs TResult when<TResult extends Object?>(TResult Function( String id,  String name,  String description)  $default,) {final _that = this;
+switch (_that) {
+case _CatalogItemDto():
+return $default(_that.id,_that.name,_that.description);case _:
+  throw StateError('Unexpected subclass');
+
+}
+}
+/// A variant of `when` that fallback to returning `null`
+///
+/// It is equivalent to doing:
+/// ```dart
+/// switch (sealedClass) {
+///   case Subclass(:final field):
+///     return ...;
+///   case _:
+///     return null;
+/// }
+/// ```
+
+@optionalTypeArgs TResult? whenOrNull<TResult extends Object?>(TResult? Function( String id,  String name,  String description)?  $default,) {final _that = this;
+switch (_that) {
+case _CatalogItemDto() when $default != null:
+return $default(_that.id,_that.name,_that.description);case _:
+  return null;
+
+}
+}
+
+}
+
+/// @nodoc
+@JsonSerializable()
+
+class _CatalogItemDto implements CatalogItemDto {
+  const _CatalogItemDto({required this.id, required this.name, required this.description});
+  factory _CatalogItemDto.fromJson(Map<String, dynamic> json) => _$CatalogItemDtoFromJson(json);
+
+@override final  String id;
+@override final  String name;
+@override final  String description;
+
+/// Create a copy of CatalogItemDto
+/// with the given fields replaced by the non-null parameter values.
+@override @JsonKey(includeFromJson: false, includeToJson: false)
+@pragma('vm:prefer-inline')
+_$CatalogItemDtoCopyWith<_CatalogItemDto> get copyWith => __$CatalogItemDtoCopyWithImpl<_CatalogItemDto>(this, _$identity);
+
+@override
+Map<String, dynamic> toJson() {
+  return _$CatalogItemDtoToJson(this, );
+}
+
+@override
+bool operator ==(Object other) {
+  return identical(this, other) || (other.runtimeType == runtimeType&&other is _CatalogItemDto&&(identical(other.id, id) || other.id == id)&&(identical(other.name, name) || other.name == name)&&(identical(other.description, description) || other.description == description));
+}
+
+@JsonKey(includeFromJson: false, includeToJson: false)
+@override
+int get hashCode => Object.hash(runtimeType,id,name,description);
+
+@override
+String toString() {
+  return 'CatalogItemDto(id: $id, name: $name, description: $description)';
+}
+
+
+}
+
+/// @nodoc
+abstract mixin class _$CatalogItemDtoCopyWith<$Res> implements $CatalogItemDtoCopyWith<$Res> {
+  factory _$CatalogItemDtoCopyWith(_CatalogItemDto value, $Res Function(_CatalogItemDto) _then) = __$CatalogItemDtoCopyWithImpl;
+@override @useResult
+$Res call({
+ String id, String name, String description
+});
+
+
+
+
+}
+/// @nodoc
+class __$CatalogItemDtoCopyWithImpl<$Res>
+    implements _$CatalogItemDtoCopyWith<$Res> {
+  __$CatalogItemDtoCopyWithImpl(this._self, this._then);
+
+  final _CatalogItemDto _self;
+  final $Res Function(_CatalogItemDto) _then;
+
+/// Create a copy of CatalogItemDto
+/// with the given fields replaced by the non-null parameter values.
+@override @pragma('vm:prefer-inline') $Res call({Object? id = null,Object? name = null,Object? description = null,}) {
+  return _then(_CatalogItemDto(
+id: null == id ? _self.id : id // ignore: cast_nullable_to_non_nullable
+as String,name: null == name ? _self.name : name // ignore: cast_nullable_to_non_nullable
+as String,description: null == description ? _self.description : description // ignore: cast_nullable_to_non_nullable
+as String,
+  ));
+}
+
+
+}
+
+// dart format on