+++ packages/api_client/lib/src/models/catalog_item_dto.g.dart
@@ -1,0 +1,21 @@
+// GENERATED CODE - DO NOT MODIFY BY HAND
+
+part of 'catalog_item_dto.dart';
+
+// **************************************************************************
+// JsonSerializableGenerator
+// **************************************************************************
+
+_CatalogItemDto _$CatalogItemDtoFromJson(Map<String, dynamic> json) =>
+    _CatalogItemDto(
+      id: json['id'] as String,
+      name: json['name'] as String,
+      description: json['description'] as String,
+    );
+
+Map<String, dynamic> _$CatalogItemDtoToJson(_CatalogItemDto instance) =>
+    <String, dynamic>{
+      'id': instance.id,
+      'name': instance.name,
+      'description': instance.description,
+    };